refactor(frontend): extract ConnectionLayer to js/graph/connection-layer.js and import in app.js
